+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Added enqueue utilities (Assets\Enqueue)
 - Added inline script and variables utilities (Assets\InlineScript, Assets\ScriptVariables, Assets\FrontendVariables, Assets\EditorVariables, Assets\AdminVariables, Assets\CriticalVariables)
+- Added utility to add plugin action links on the Plugins table page (Plugin\ActionLinks)
+- Added utility to add plugin settings link on the Plugins table page (Plugin\SettingsLink)
 
 ## [0.1.0] - 2023-04-06